Implement generic multi-channel confocal handling and JSON default setup parameters
Implement generic multi-channel confocal handling, that reads the number of confocal images from the Imspector stack, and automatically generates an image vector that handles all confocal images, and inputs them to analysis pipeline (as many as possible).
Add JSON default setup parameters, which can be set by the user before starting etMINFLUX, and from which values are read and automatically set in the GUI when etMINFLUX is started. Remove some of these values, such as timing values, from the GUI, as they should not be controllable from there.